+++ README.md
@@ -1,2 +1,29 @@
-# owl-partners-web
-Frontend do projeto "Owl Partners" (5º DSM - 2023, FATEC Profº Jessen Vidal - SJC)
+<h1 align="center"> Projeto "Owl Partners" </h1>
+
+Este repositório se refere ao projeto "Owl Partners", que consiste em um sistema para gerenciamento de parcerias, onde administradores cadastram parcerias em potencial e atualizam informações acerca das reuniões ocorridas, assim monitorando o status em que a instituição se encontra no processo de fechamento de parceria.
+
+> Aplicação desenvolvida por alunos do 5º semestre do tecnólogo em Desenvolvimento de Software Multiplataforma, na FATEC Profº Jessen Vidal - São José dos Campos, SP :rocket:
+
+### :hammer_and_wrench: Tecnologias
+
+As seguintes tecnologias e ferramentas foram utilizadas neste projeto: `Figma, React, TypeScript, HTML, CSS, Axios, Styled Components`
+
+### :gear: Como utilizar
+
+É possível acessar a aplicação pelo link de seu deploy (clicando [aqui](owl-partners.netlify.app/auth)) mas também é possível rodar localmente seguindo o passo a passo abaixo:
+
+- Tutorial para rodar o projeto
+
+Com o [Node](https://nodejs.org/en/) instalado em sua máquina, utilize estes comandos em um terminal:
+
+```bash
+# Baixe ou clone este repositório
+$ git clone https://github.com/The-Bugger-Ducks/owl-partners-web.git
+
+# Acesse a pasta, instale as dependências e inicie o projeto
+$ cd owl-partners-web
+$ npm i
+$ npm start
+```
+
+A aplicação inciará localmente na porta 3000. Use o navegador para acessar o link [http://localhost:3000](http://localhost:3000) e executar as funcionalidades da aplicação.
