--- owl-partners/src/pages/Auth/Auth.tsx
@@ -1,3 +1,0 @@
-export default function Auth() {
-    return <>Olá, mundo 🌏 Auth</>
-}
