--- owl-partners/src/pages/Dashboard/Dashboard.tsx
@@ -1,3 +1,0 @@
-export default function Dashboard() {
-    return <>Olá, mundo 🌏 Dashboard</>
-}
