--- owl-partners/src/utils/enums/endpoints.ts
@@ -1,6 +1,0 @@
-enum Endpoints {
-    DASHBOARD = "/dashboard",
-    AUTH = "/auth"
-}
-
-export default Endpoints
