--- owl-partners/src/utils/enums/userRole.ts
@@ -1,6 +1,0 @@
-enum UserRole {
-    ADMIN = "ADMIN",
-    SIMPLE = "SIMPLE",
-}
-
-export default UserRole
