--- owl-partners/src/utils/interfaces/user.ts
@@ -1,12 +1,0 @@
-import UserRole from "../enums/userRole";
-
-interface User {
-    id: string;
-    name: string;
-    lastName: string;
-    email: string;
-    role: UserRole;
-    createdAt: Date;
-}
-
-export default User;
