+++ package.json
@@ -8,7 +8,6 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.27",
-    "@types/react": "^18.2.6",
     "@types/react-dom": "^18.2.4",
     "axios": "^1.4.0",
     "react": "^18.2.0",
@@ -42,6 +41,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react": "^18.2.6",
     "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.59.5",
     "eslint": "^8.40.0",
