+++ src/App.tsx
@@ -1,6 +1,6 @@
-import React from 'react';  
-import MyRoutes from './routes';
-import { BrowserRouter } from 'react-router-dom';
+import React from "react";
+import MyRoutes from "./routes";
+import { BrowserRouter } from "react-router-dom";
 
 function App() {
   return (
