+++ src/pages/Auth/Auth.tsx
@@ -1,0 +1,3 @@
+export default function Auth() {
+  return <>Olá, mundo 🌏 Auth</>;
+}
