+++ src/pages/Dashboard/Dashboard.tsx
@@ -1,0 +1,3 @@
+export default function Dashboard() {
+  return <>Olá, mundo 🌏 Dashboard</>;
+}
