+++ src/routes.tsx
@@ -1,0 +1,13 @@
+import { Routes, Route } from "react-router-dom";
+
+import Dashboard from "./pages/Dashboard/Dashboard";
+import Auth from "./pages/Auth/Auth";
+
+export default function MyRoutes() {
+  return (
+    <Routes>
+      <Route path="/dashboard" element={<Dashboard />} />
+      <Route path="/" element={<Auth />} />
+    </Routes>
+  );
+}
