+++ src/utils/enums/endpoints.ts
@@ -1,0 +1,6 @@
+enum Endpoints {
+  DASHBOARD = "/dashboard",
+  AUTH = "/auth",
+}
+
+export default Endpoints;
