+++ src/utils/enums/userRole.ts
@@ -1,0 +1,6 @@
+enum UserRole {
+  ADMIN = "ADMIN",
+  SIMPLE = "SIMPLE",
+}
+
+export default UserRole;
